Use a single rollup config
