+++ packages/rpv-default-layout/package.json
@@ -43,7 +43,7 @@
     },
     "scripts": {
         "clean": "rm -rf npm && mkdir npm",
-        "build": "npm run clean && cp src/index.d.ts npm && cp dist/index.js npm && rollup -c && npm run mincss && rm -rf npm/cjs/rpv-default-layout.less",
+        "build": "npm run clean && cp src/index.d.ts npm && cp dist/index.js npm && rollup -c ../../rollup.config.js && npm run mincss && rm -rf npm/cjs/rpv-default-layout.less",
         "mincss": "postcss npm/cjs/rpv-default-layout.css > npm/cjs/rpv-default-layout.min.css"
     }
 }
