+++ packages/rpv-full-screen/package.json
@@ -39,7 +39,7 @@
     },
     "scripts": {
         "clean": "rm -rf npm && mkdir npm",
-        "build": "npm run clean && cp src/index.d.ts npm && cp dist/index.js npm && rollup -c && npm run mincss && rm -rf npm/cjs/rpv-full-screen.less",
+        "build": "npm run clean && cp src/index.d.ts npm && cp dist/index.js npm && rollup -c ../../rollup.config.js && npm run mincss && rm -rf npm/cjs/rpv-full-screen.less",
         "mincss": "postcss npm/cjs/rpv-full-screen.css > npm/cjs/rpv-full-screen.min.css"
     }
 }
