--- packages/rpv-open/rollup.config.js
@@ -1,49 +1,0 @@
-import less from 'rollup-plugin-less-modules';
-import { terser } from 'rollup-plugin-terser';
-import typescript from '@rollup/plugin-typescript';
-
-const pkg = require('./package.json');
-
-const external = [
-    ...Object.keys(pkg.dependencies || {}),
-    ...Object.keys(pkg.peerDependencies || {}),
-];
-
-const input = './src/index.ts';
-
-export default [
-    // CJS
-    {
-        input,
-        output: {
-            file: './npm/cjs/rpv-open.js',
-            format: 'cjs',
-        },
-        external,
-        plugins: [
-            less({
-                output: './npm/cjs/rpv-open.css',
-                sourcemap: false,
-            }),
-            typescript(),
-        ],
-    },
-
-    // Minified CJS
-    {
-        input,
-        output: {
-            file: './npm/cjs/rpv-open.min.js',
-            format: 'cjs',
-        },
-        external,
-        plugins: [
-            less({
-                output: './npm/cjs/rpv-open.css',
-                sourcemap: false,
-            }),
-            typescript(),
-            terser(),
-        ],
-    }
-];
