+++ packages/rpv-page-navigation/package.json
@@ -39,7 +39,7 @@
     },
     "scripts": {
         "clean": "rm -rf npm && mkdir npm",
-        "build": "npm run clean && cp src/index.d.ts npm && cp dist/index.js npm && rollup -c && npm run mincss && rm -rf npm/cjs/rpv-page-navigation.less",
+        "build": "npm run clean && cp src/index.d.ts npm && cp dist/index.js npm && rollup -c ../../rollup.config.js && npm run mincss && rm -rf npm/cjs/rpv-page-navigation.less",
         "mincss": "postcss npm/cjs/rpv-page-navigation.css > npm/cjs/rpv-page-navigation.min.css"
     }
 }
