+++ packages/rpv-properties/package.json
@@ -39,7 +39,7 @@
     },
     "scripts": {
         "clean": "rm -rf npm && mkdir npm",
-        "build": "npm run clean && cp src/index.d.ts npm && cp dist/index.js npm && rollup -c && npm run mincss && rm -rf npm/cjs/rpv-properties.less",
+        "build": "npm run clean && cp src/index.d.ts npm && cp dist/index.js npm && rollup -c ../../rollup.config.js && npm run mincss && rm -rf npm/cjs/rpv-properties.less",
         "mincss": "postcss npm/cjs/rpv-properties.css > npm/cjs/rpv-properties.min.css"
     }
 }
