+++ packages/rpv/package.json
@@ -34,7 +34,7 @@
     ],
     "scripts": {
         "clean": "rm -rf npm && mkdir npm",
-        "build": "npm run clean && cp src/index.d.ts npm && cp dist/index.js npm && rollup -c && npm run mincss && rm -rf npm/cjs/rpv.less",
+        "build": "npm run clean && cp src/index.d.ts npm && cp dist/index.js npm && rollup -c ../../rollup.config.js && npm run mincss && rm -rf npm/cjs/rpv.less",
         "mincss": "postcss npm/cjs/rpv.css > npm/cjs/rpv.min.css"
     },
     "dependencies": {
